+++ src/Model/Command/UserRecommendation.php
@@ -29,18 +29,27 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
     private $hardRotation = false;
     /** @var string */
     private $minimalRelevance = self::MINIMAL_RELEVANCE_LOW;
-    /** @var array */
+    /** @var string[] */
     private $filters = ['valid_to >= NOW'];
     /** @var string|null */
     private $modelName = null;
+    /** @var string[] */
+    private $responseProperties = [];
 
-    private function __construct(string $userId, int $count, string $scenario, float $rotationRate, int $rotationTime)
-    {
+    private function __construct(
+        string $userId,
+        int $count,
+        string $scenario,
+        float $rotationRate,
+        int $rotationTime,
+        array $responseProperties
+    ) {
         $this->setUserId($userId);
         $this->setCount($count);
         $this->setScenario($scenario);
         $this->setRotationRate($rotationRate);
         $this->setRotationTime($rotationTime);
+        $this->setResponseProperties($responseProperties);
     }
 
     /**
@@ -53,6 +62,7 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
      * Set from 0.0 for no rotation (same items will be recommended) up to 1.0 (same items should not be recommended).
      * @param int $rotationTime Specify for how long will the item's rotationRate be taken in account and so the item
      * is penalized for recommendations.
+     * @param string[] $responseProperties Specify which properties you want to retrieve from Matej alongside the item_id.
      * @return static
      */
     public static function create(
@@ -60,9 +70,10 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
         int $count,
         string $scenario,
         float $rotationRate,
-        int $rotationTime
+        int $rotationTime,
+        array $responseProperties = []
     ): self {
-        return new static($userId, $count, $scenario, $rotationRate, $rotationTime);
+        return new static($userId, $count, $scenario, $rotationRate, $rotationTime, $responseProperties);
     }
 
     /**
@@ -119,6 +130,33 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
         Assertion::allString($filters);
 
         $this->filters = $filters;
+
+        return $this;
+    }
+
+    /**
+     * Add another response property you want returned. item_id is always returned by Matej.
+     */
+    public function addResponseProperty(string $property): self
+    {
+        Assertion::typeIdentifier($property);
+
+        $this->responseProperties[] = $property;
+
+        return $this;
+    }
+
+    /**
+     * Set all response properties you want returned. item_id is always returned by Matej, even when you don't specify it.
+     *
+     * @param string[] $properties
+     * @return $this
+     */
+    public function setResponseProperties(array $properties): self
+    {
+        Assertion::allTypeIdentifier($properties);
+
+        $this->responseProperties = $properties;
 
         return $this;
     }
@@ -198,6 +236,7 @@ class UserRecommendation extends AbstractCommand implements UserAwareInterface
             'hard_rotation' => $this->hardRotation,
             'min_relevance' => $this->minimalRelevance,
             'filter' => $this->assembleFiltersString(),
+            'properties' => $this->responseProperties,
         ];
 
         if ($this->modelName !== null) {